INVENTORY/inventory_system/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.4.0
pillow==11.3.0
pyserial==3.5
reportlab==4.4.3
SQLAlchemy==2.0.41
tk==0.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
